+++ lang/en/mposter.php
@@ -17,7 +17,7 @@
 /**
  * Strings of the Media Poster activity module are defined here.
  *
- * @package     mod_mdposter
+ * @package     mod_mposter
  * @category    string
  * @copyright   2015 David Mudrak <david@moodle.com>
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
@@ -45,28 +45,28 @@ $string['meta_value']  = 'Value';
 $string['meta_title']  = 'Title';
 $string['modulename_help'] = 'The Media Poster module enables a teacher to make a page out of blocks (such as the HTML block, comments, calendar or RSS feed block), dragging and dropping them to the preferred location.
 
-You can use the mdposter
+You can use the mposter
 
 * As a space for your students to present results of their work.
 * As a course dashboard, collating useful elements for students on a single page and keeping the main course page clean and clear of side blocks.';
 $string['modulenameplural']     	= 'Media Posters';
-$string['page-mod-mdposter-view'] 	= 'Media Poster module main page';
+$string['page-mod-mposter-view'] 	= 'Media Poster module main page';
 $string['pluginadministration'] 	= 'Media Poster administration';
 $string['pluginname'] 				= 'Media Poster';
-$string['mdposter:addinstance'] 		= 'Add a new mdposter';
-$string['mdposter:view'] 				= 'View mdposter';
-$string['mdpostername'] 				= 'Title';
+$string['mposter:addinstance'] 		= 'Add a new mposter';
+$string['mposter:view'] 				= 'View mposter';
+$string['mpostername'] 				= 'Title';
 $string['overwrite'] 				= 'Overwrite Intermusic metadata with custom fields.';
-$string['mdposterauthor'] 			= 'Composer';
-$string['mdpostersurtitle'] 			= 'Surtitle';
-$string['mdposternumbering'] 			= 'Numbering';
-$string['mdposterlanguage'] 			= 'Language';
+$string['mposterauthor'] 			= 'Composer';
+$string['mpostersurtitle'] 			= 'Surtitle';
+$string['mposternumbering'] 			= 'Numbering';
+$string['mposterlanguage'] 			= 'Language';
 $string['autopopulate'] 			= 'Autopopulate from childern \'File\' blocks';
 $string['privacy:metadata'] 		= 'Media Poster does not store any personal data';
 $string['showdescriptionview'] 		= 'Display description on view page';
-$string['showdescriptionview_help'] = 'If enabled, the description above will be displayed on the mdposter page.';
+$string['showdescriptionview_help'] = 'If enabled, the description above will be displayed on the mposter page.';
 $string['shownameview'] 			= 'Display name on view page';
-$string['shownameview_help'] 		= 'If enabled, the mdposter name will be displayed as heading on the mdposter page.';
+$string['shownameview_help'] 		= 'If enabled, the mposter name will be displayed as heading on the mposter page.';
 $string['metadatafile']    = 'Metadata file';
 $string['resourcecontent'] = 'Files and subfolders';
 $string['metadata_task'] = 'Update Metadata';