+++ CHANGES.md
@@ -17,11 +17,11 @@
 
 * In editing mode, the "Add a block" drop down selector is now displayed
   in the Boost based themes, too - making it easier and more intuitive
-  to add blocks to the mdposter regions.
+  to add blocks to the mposter regions.
 
 ### 5.0.0 ###
 
-* Improved the styling of the mdposter editing page
+* Improved the styling of the mposter editing page
 * Tested with Moodle 3.2 and 3.3 under Boost, Clean and More themes
 
 ### 4.0.1 ###
@@ -41,7 +41,7 @@
 
 ### v2 ###
 
-* Fixed the missing module help string (displayed when adding the mdposter into the course).
+* Fixed the missing module help string (displayed when adding the mposter into the course).
 
 ### v1 ###